code/requirements.txt
absl-py==0.15.0
AST==0.0.2
astor==0.8.1
backports-abc==0.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
certifi==2020.6.20
cloudpickle==0.5.3
configparser==4.0.2
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==5.1.0
enum34==1.1.10
funcsigs==1.0.2
future==0.18.2
futures==3.4.0
gast==0.5.3
grpcio==1.41.1
importlib-metadata==2.1.3
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Markdown==3.1.1
matplotlib==1.5.1
mock==3.0.5
numpy==1.16.1
opencv-python==3.4.1.15
path.py==11.5.2
pathlib2==2.3.7.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==3.1.2
pip==19.3.1
prompt-toolkit==1.0.15
protobuf==3.17.3
psutil==3.4.2
ptyprocess==0.7.0
Pygments==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2022.6
pyzmq==18.1.0
scandir==1.10.0
scipy==1.2.3
setuptools==44.0.0.post20200106
simplegeneric==0.8.1
singledispatch==3.7.0
six==1.16.0
tensorboard==1.10.0
tensorflow==1.8.0
termcolor==1.1.0
tornado==5.1.1
traitlets==4.3.3
typing==3.10.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.37.1
zipp==1.2.0
